[PVR] CDirectoryProvider: Now supports async PVR startup. Fixes favorite tv and radio channels not appearing on Home screen.
