+++ xbmc/listproviders/DirectoryProvider.cpp
@@ -33,6 +33,7 @@
 #include "music/dialogs/GUIDialogMusicInfo.h"
 #include "music/MusicThumbLoader.h"
 #include "pictures/PictureThumbLoader.h"
+#include "pvr/PVRManager.h"
 #include "settings/Settings.h"
 #include "threads/SingleLock.h"
 #include "utils/JobManager.h"
@@ -48,6 +49,7 @@
 using namespace XFILE;
 using namespace ANNOUNCEMENT;
 using namespace KODI::MESSAGING;
+using namespace PVR;
 
 class CDirectoryJob : public CJob
 {
@@ -263,7 +265,7 @@ void CDirectoryProvider::Fetch(std::vector<CGUIListItemPtr> &items) const
   }
 }
 
-void CDirectoryProvider::OnEvent(const ADDON::AddonEvent& event)
+void CDirectoryProvider::OnAddonEvent(const ADDON::AddonEvent& event)
 {
   CSingleLock lock(m_section);
   if (URIUtils::IsProtocol(m_currentUrl, "addons"))
@@ -272,7 +274,20 @@ void CDirectoryProvider::OnEvent(const ADDON::AddonEvent& event)
         typeid(event) == typeid(ADDON::AddonEvents::Disabled) ||
         typeid(event) == typeid(ADDON::AddonEvents::InstalledChanged) ||
         typeid(event) == typeid(ADDON::AddonEvents::MetadataChanged))
-    m_updateState = INVALIDATED;
+      m_updateState = INVALIDATED;
+  }
+}
+
+void CDirectoryProvider::OnPVRManagerEvent(const PVR::ManagerState& event)
+{
+  CSingleLock lock(m_section);
+  if (URIUtils::IsProtocol(m_currentUrl, "pvr"))
+  {
+    if (event == ManagerStateStarted ||
+        event == ManagerStateStopped ||
+        event == ManagerStateError ||
+        event == ManagerStateInterrupted)
+      m_updateState = INVALIDATED;
   }
 }
 
@@ -300,6 +315,7 @@ void CDirectoryProvider::Reset(bool immediately /* = false */)
       m_isAnnounced = false;
       CAnnouncementManager::GetInstance().RemoveAnnouncer(this);
       ADDON::CAddonMgr::GetInstance().Events().Unsubscribe(this);
+      g_PVRManager.Events().Unsubscribe(this);
     }
   }
 }
@@ -393,7 +409,8 @@ bool CDirectoryProvider::UpdateURL()
   {
     m_isAnnounced = true;
     CAnnouncementManager::GetInstance().AddAnnouncer(this);
-    ADDON::CAddonMgr::GetInstance().Events().Subscribe(this, &CDirectoryProvider::OnEvent);
+    ADDON::CAddonMgr::GetInstance().Events().Subscribe(this, &CDirectoryProvider::OnAddonEvent);
+    g_PVRManager.Events().Subscribe(this, &CDirectoryProvider::OnPVRManagerEvent);
   }
   return true;
 }
